refactor: standardize header casing
