+++ include/game_client/mouse.h
@@ -54,7 +54,7 @@
 // USER INCLUDES //////////////////////////////////////////////////////////////
 #include "Lib/BaseType.h"
 #include "GameEngine/Common/subsystem_interface.h"
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 #include "Common/UnicodeString.h"
 
 // FORWARD REFERENCES /////////////////////////////////////////////////////////
