+++ include/game_engine/common/ascii_string.h
