+++ include/game_engine/common/directory.h
@@ -33,7 +33,7 @@
 #define __DIRECTORY_H__
 
 // INCLUDES ///////////////////////////////////////////////////////////////////////////////////////
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 #include "Common/STLTypedefs.h"
 
 // FORWARD REFERENCES /////////////////////////////////////////////////////////////////////////////
