+++ include/game_engine/common/ini.h
@@ -36,7 +36,7 @@
 #include <stddef.h>	// for offsetof, which we don't use but everyone who includes us does
 #include "Common/NameKeyGenerator.h"
 #include "Common/STLTypedefs.h"
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 #include "Common/Science.h"
 #include "Common/GameCommon.h"
