+++ include/game_engine/common/snapshot.h
@@ -34,7 +34,7 @@
 #define __SNAPSHOT_H_
 
 // USER INCLUDES //////////////////////////////////////////////////////////////////////////////////
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 
 // FORWARD REFERENCES /////////////////////////////////////////////////////////////////////////////
 class Xfer;
