+++ include/game_engine_device/w3d_device/game_client/terraintex.h
@@ -35,7 +35,7 @@
 
 #include "WW3D2/Texture.h"
 #include "WWMATH/Matrix3d.h"
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 
 class WorldHeightMap;
 #define TILE_OFFSET 8
