+++ include/game_engine_device/w3d_device/game_client/w3dbridgebuffer.h
@@ -60,7 +60,7 @@
 #include "Lib/BaseType.h"
 #include "common/GameType.h"
 #include "common/Dict.h"
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 
 //-----------------------------------------------------------------------------
 //           Forward References
