+++ include/game_engine_device/w3d_device/game_client/w3dcustomedging.h
@@ -59,7 +59,7 @@
 #include "vertmaterial.h"
 #include "Lib/BaseType.h"
 #include "common/GameType.h"
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 
 //-----------------------------------------------------------------------------
 //           Forward References
