+++ include/game_engine_device/w3d_device/game_client/w3dpropbuffer.h
@@ -55,7 +55,7 @@
 #include "w3d_file.h"
 #include "Lib/BaseType.h"
 #include "common/GameType.h"
-#include "common/asciiString.h"
+#include "common/ascii_string.h"
 #include "common/GlobalData.h"
 
 //-----------------------------------------------------------------------------
